Retry sheets requests if we've maxed out on reqs per minute.
Since there is apparently a limit of 60 read or write requests per minute to google sheets, if we get back the http error that indicates our limit, wait 60 seconds and keep going. To minimize code, all the sheets methods are now private, and wrapped in a public method wrapped in a try-except block.